Don't use human attitude value in AI code
[change id advc.130u]

Added an assertion to the AI attitude functions that warns when the attitude of a human player is accessed. Can pass bAssertNonHuman=false as a parameter to bypass the assertion.

It's OK to use the human attitude for updating other AI values (e.g. strategy hash, worst enemy) that, ultimately, remain unused unless AI Auto Play is activated (or when a player retires from a multiplayer game). Actually, attitude and worst enemy get updated in those cases anyway, and the strategy hash could be (is already?) updated as well, but it's nice to have all the cached AI data available for inspection in Debug mode.

An earlier change had already treated humans as being Cautious toward everyone, but I think it's better to address human players explicitly whenever the AI takes the attitude value of a rival into account.

Fixed several uses of human attitude in the UWAI code, one in change [advc.109] and one in the K-Mod computation of espionage weights. [advc.120]

Replaced some hardcoded thresholds for relations values in AI code with thresholds computed from the CvGlobals::RELATIONS_THRESH_... values loaded from XML.

And a few unrelated changes to comments.