Don't obscure memory-based relations when memory>1
Amendment to this commit:
SHA 714de64ef00815f320a45f859c6f009fcb82a791

Only concerns the random personalities option. When e.g. a leader with the Gandhi personality remembers multiple wars declared on his friends, then the displayed relations modifier should not be just -1. For simplicity, I just show the true relations modifier in such cases, i.e. -4 in the example. (In the code, I'm checking if the memory count is less than or equal to 2 because AdvCiv counts memory on a times-2 scale.)