Remove non-human assertion from AI_getAttitude
[change id advc.130u]

I had added that in a recent commit along with parameters for ignoring the assertion. Seems to be more trouble than it's worth. I've now instead improved the computation of the attitude value assumed for human players by the AI: Friendly for teammates of the human player, Annoyed for human rivals of the human player and either Catious or Annoyed for AI rivals of the human player depending on the AI attitude toward the human player. Previously, it was Cautious in all those cases.

This is probably better than most of the custom code that I use at some of the AI_getAttitude call locations, but it's difficult to track that code down (there are a lot of call locations), so I'll remove it as I happen to come across it.

Attitudes within a team are now always treated as Friendly, even if the bForce parameter is set. That parameter only affects relationships of vassals and masters now.