CvInfoEnum class and example added
I'm not yet sure if this change is worthwhile. The goal is to make it easier to implement XML schema changes on the DLL side and – later maybe – to reduce code duplication in the read(CvXMLLoadUtility*) functions of classes (currently) derived from CvInfoBase.

Elements added in this way are accessed as follows (example):
kImprovement.get(CvImprovementInfo::DefenseModifier)
instead of kImprovement.getDefenseModifier()
where kImprovement is a CvImprovementInfo reference.

Ideally, to add a new XML element tag, the name of the tag would have to be specified only once in the DLL. This is possible through macro definitions and that works pretty well for GlobalDefines (change advc.003t; see CvGlobals.h) but requires too much error-prone code to be added to each info class in this case.

Note that a separate enum type is needed for each data type supported by XML. Otherwise, the get function couldn't be overloaded and uglier names such as getInt, getBool would have to be used. So far I've only added enum types for int and bool. To be done maybe: float, string, integer pair (e.g. TerrainMakesValid), integer triple (e.g. RouteYieldChanges). For the latter two, memory usage could then be optimized.

The next best thing, if specifying the tag name once isn't enough, would be to specify it as an enum value (e.g. DefenseModifier) and as a string ("DefenseModifier") side by side in one place. Can't do that either because the enum value has to be defined in a header file and the string representation doesn't belong in a header file. Moreover, it needs to be possible to specify a default value as well. It would be much better to let the schema file set a default value, but I'm not sure how much of the code for parsing schema files is accessible to the DLL. As it is, the tag data (enum value, string, default value) needs to be wrapped in a small "XMLElement" object. Thus, the tag name has to be specified once in the header file and twice (in one place) in the implementation file. This is still better than the status quo ante when it had to be added in 4 different places, 5 if not inlined (plus possibly another 2 for the XML cache).

I haven't found a way to expose my get functions to Python. That means, if an XML element added through an enum value is supposed to be exposed to Python, a wrapper function will have to be added.

Note that generating the getter functions from the schema file would only be possible through an external code generator – through C++ alone, we'd end up with something inefficient like the CvGlobals::getDefineINT(char const*) function. Nightinggale has worked on a Perl script (https://github.com/We-the-People-civ4col-mod/Mod/commit/52780d861660468ed3a5bdae22ef38954ac6d1be ), but I think that only deals with the "type" elements in XML, i.e. it doesn't generate any info accessor functions.

As for the new "CvInfoEnum" class: It represents XML element tags as enum values. That's probably not clear from the class name; don't want a long class name though because derived classes need to refer to some base members explicitly. I've tried implementing the enum handling directly at CvInfoBase, but that resulted in a crash in the EXE. My guess is that there is some class derived from CvInfoBase (probably a small one) whose size mustn't change. Well, it's also neater to put the enum code in a new class.

[change id: advc.tag]