Disable assertions in preAutoSave, preQuickSave
I had added these as part of change [advc.106l]. They're in principle helpful for detecting failed CvEngine::AutoSave calls, but that function also "fails" when no save is created due to the AutoSaveInterval setting in CivilizationIV.ini. I.e. the assertions only work correctly when the interval is 1.

And some minor changes to comments.