Fix an issue with AI workers not improving small islands
A recent change [id advc.113] to the AI computation of the number of workers needed on a landmass had resulted in 0 workers for islands without cities, blocking change [advc.040] which is supposed to let the AI improve islands in the radius of a mainland city.

And a bugfix for a smaller issue with [advc.107]. Part of the AI computation of the number of defenders needed in a colony was on the wrong scale.