Don't check array bounds when looping over all plots
Replaced most calls to CvMap::plotByIndex with a new function getPlotByIndex, which returns a CvPlot reference and doesn't check array bounds. When looping over all plots on the map (from index 0 to CvMap::numPlots()), there is no danger of going out of bounds. Some of those loops are relevant for overall performance, so it's nice to get rid of unnecessary branching.

As a side-effect, most of the loops over all plots now use CvPlot references, and, to accommodate that, I've changed the signatures of some functions that take a CvPlot parameter. (Eventually, const& should become the default for passing objects, but, so far, pointers are still predominant.)

I had made a similar change for the CvMap::plot function last month:
SHA 22e3de2096b84264109fddd7963632d2def71447