Use AgentIterators in all UWAI classes
... replacing the local player and team sequence caches used previously.
Also converted UWAI cache to use EnumMap, switched to stdext::hash_map and inlined some functions.
Refactored MilitaryAnalyst, InvasionGraph so that a flat_set can easily be tested (to be done) in place of std::set; also improved type safety and safety of array and vector accesses in those classes.