Field-of-view slider tweaks
[change id advc.090]

Slightly longer bar (at least on high resolutions), which makes it a bit easier to set the slider to a particular value. However, the issue described here (at the end of the post)
https://forums.civfanatics.com/threads/advanced-civ.614217/page-24#post-15610934
isn't (and probably won't be) fixed.

The text labels show a percentage (10 to 100%) instead of the true FoV value. From 10 to 75%, the percentage is equal to the true FoV value, above 75%, the slider becomes more sensitive, which should allow Huge maps to be fully displayed (e.g. for AI Auto Play or a screenshot).