Prohibit comparisons between (certain) enum types
By declaring overloaded comparison operators without a definition for pairs of enum types that are easy to confuse, especially PlayerTypes and TeamTypes. [advc.enum]

Found and fixed a bunch of errors this way - all in AdvCiv code. (There used to be such errors also in BtS code, but it seems that modders have already fixed all of them over the years.)