Update player AI cache data after loading savegame
The cache in question contains some misc. CvPlayerAI data and was introduced by K-Mod. Updating the cache after loading means that the cache data don't necessarily have to be serialized. The cache normally gets updated at the start of a player's turn. That means, only the cache of the active human player didn't get updated after loading. The human cache matters for testing through AI Auto Play, so, for development purposes this is an important bugfix. (All the K-Mod cache data was being serialized, so the error had been mine alone.)