Hover text for tiles that cannot be entered
When hovering for combat odds and the only (otherwise) valid defenders are maximally damaged, the help text says so and states the attacker's damage limit. [advc.089]

When hovering for movement into terrain that one of the selected units cannot enter, help text says which technology if any would allow the move or if being the owner of the tile would help. [advc.057]

The latter change is intended for mod-mods (had been requested by Cruiser76). In AdvCiv, Astronomy is the only tech that allows a unit to enter otherwise impassable terrain (Work Boat, Ocean).