Revised AI calculation of culture garrisons
More accurate estimate of the number of units needed in a city in order to bring the revolt chance down to an acceptable risk. The per-area target for the number of floating defenders is increased a bit based on culture pressure. In cases that look hopeless, no extra defenders are allotted. [advc.099c]

(Not sure how quickly the AI will be able to reinforce cities that are about to revolt. AI units in non-defensive roles still don't respond to imminent revolts. The AI also still makes no use of the revolt protection from the Leadership promotion.)