Adjust/ fix decay probabilities of AI diplo memory counters
Changed the formula for faster decay when multiple diplo actions are remembered. For example, two declarations of war should now be fully forgotten after an expected duration of 150+75 turns on Normal speed. Three declarations after 150+75+50 turns. Etc. My old formula was ... more complicated and incorrectly implemented. [advc.130r]

Memory about tech trades ("you're becoming too advanced", "you shared your discoveries") are exempt from the above, i.e. they decay as in BtS with a per-turn probability that doesn't depend on the memory count. However, for "you shared your discoveries", I've increased the per-turn probability by a factor of 4, meaning that one tech is forgotten, on average, every 25 turns instead of every 100 turns. The relations modifier per tech is now between 0.08 and 0.2 (was between 0.05 and 0.2) and fractions get rounded to the nearest integer. (The rounding change is not part of this commit; has been in place since v0.8 or so. BtS rounds down.) On the bottom line, +1 from shared discoveries should be easier to get than in BtS, but more difficult to sustain. (Not sure about higher modifiers.) Still to be tested. [advc.553]

"You gave us help" now decays with a (base) probability of 1/150; was 1/200. [advc.144]