Help text for GW protection ability; fix recent Civilopedia bug.
I'm appending "and protects it from Global Warming" to "Gives vegetation a higher chance to spread" (that's the help text of the feature spread ability).

Right-click on a Worker Build icon on the tech tree now jumps to the correct Terrain Improvement article in Civilopedia. I had broken that a while ago when refactoring DLL-to-Python calls.