Remove global getAsyncRand function
This reverts part of SHA d844afa452d1c4a77b74bdcbfc802b89a014ab10

Since CvGlobals has only a pointer to the non-synchronized PRNG, there is no problem with const-correctness – and so that (truly) global accessor function was an unnecessary addition.

(I'm in the process of filling some gaps in the documentation part of the manual.)

Also: My general cleanup of CvGlobals is no longer tagged with "advc.003t", but just with "advc" or "advc.opt".