Tweak number-of-cities maintenance modifiers
As outlined in the second half of this CFC post
https://forums.civfanatics.com/threads/advanced-civ.614217/page-25#post-15651246
I've lowered the maintenance modifier on Huge maps a little bit and added a crowdedness modifier to account for the initial player count and sea level.
[advc.140]

The hard limit on the city count in the maintenance formula is still in place; I've merely increased it a bit. [advc.exp.1]