Fix issues with AI Auto Play ending upon player defeat
[change id 127]

When the active player is defeated during AI Auto Play, the player with the next highest id automatically becomes the active player; Auto Play continues uninterrupted.

The popup in AIAutoPlay.py is now used only as a fallback. I've fixed some issues with that, but it's still not working reliably and some assertions will fail even when it works correctly.