ScaledInt: Improve accuracy of operators, fix enum type issues.
Multiplication and division now call one of the static mulDiv functions directly instead of first converting the operand with the smaller scale to the larger scale. This leads to more accurate results and may also be faster.

The comparison operators now convert both operands to the product of the two scales (unless the scales are the same), using 64 bit if necessary. This makes comparisons exact - as they should be.

My approach to the EnumType parameter -- defining function templates only for matching types -- doesn't work after all; at least not so easily. For now, matching types are enforced by static assertions.

And I've renamed "scaled_int" to "scaled" and "scaled_uint" to "uscaled". The default scale factor is now 2048. The class template is still named ScaledInt.