Upon saving, delete popups stored by the EXE at AI players.
Not (yet) sure how this happens (and whether it also happens in BtS), but, apparently, the EXE can place popups in the popup queues of AI players that briefly appear when taking control of an AI player through Alt+Z and otherwise never get deleted. Now the popup queues of AI players are cleared before saving. [advc.001]