City trade tweaks
[change id advc.ctr]

Show a note in the city-conquered popup when the current target player for liberation differs from the target at peacetime (i.e. when liberation to the original owner is only possible through the city-conquered popup.

AI again (as in BtS) says that "no such deal is possible" when asked what it will give in exchange for liberating a city.

AI more reluctant to liberate a city for free when the recipient is the worst enemy of other relevant AI civs.

Increased trade memory for liberation:
* More lenient thresholds to determine that a liberated city is a major city or insignificant.
* A single city of medium importance results in "+2: You liberated our cities" despite rounding.
* Both sides' trade values are counted for fair trade and enemy trade memory, but still reduced a bit (because there's also liberation memory).
* Fair/ enemy trade memory is also counted when ceding a city through the city-conquered popup.