Clean up "concealed player identity" mess
[change id advc.058]

K-Mod had changed CvPlayer::getName, getPlayerColor and the two getCivDescription functions so that they return "unknown" or the Barbarian color for players whom the active (human) player hasn't met or seen yet. That has been leading to some problems now and then, and it turns out that it's not a workable approach at all for the Rise & Fall component, which collects interface messages sent to AI civs in case that the human player will later take control of those civs. The player names and colors in those messages mustn't be based on the active player. Bug report:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-32#post-15739533

Now the true names and colors are returned again by default, and there are separate functions getKnownName, getKnownPlayerColor etc. for the few situations in which identities may have to be concealed. Hopefully, I've correctly identified those situations.

And refactored CvPlayerColorInfo so that it uses ColorTypes instead of int.