Fix issues with Taj Mahal Golden Age length
The effect of buildings that start a Golden Age (GA) is now delayed until the start of the owner's next turn. [advc.001x]
(BtS/Vanilla starts the GA right away and adds one GA turn because the current turn is almost over. This leads to a full extra GA turn when the building is completed during a GA. If no GA is ongoing, then the BtS behavior is still wrong because it'll already affect production and GPP generated by other cities during the end-of-turn sequence.)