Separate flag for normalization in CitySiteEvaluator
When placing starting sites and when deciding how to normalize starting sites, found-city values need to be computed with some special provisions; not quite the same in both cases. BtS had a flag 'bStartingLoc' and normalization was determined based on that and by checking if starting locations had already been assigned. Meaning that, by default, the bStartingLoc behavior had applied also when normalizing. [advc.031e]
(I think it's better to be explicit about this and I may want to be able to compute found values as if normalizing as part of a fairer algorithm for placing starting sites. [advc.027])

Goody huts in the starting city radius are now counted for normalization. (It's a start ...) [advc.031]