Show map rand, sync rand seed in Debug mode
[change id advc.027b]

That is, store the initial seed, and show it in Debug mode if Ctrl is held down when hovering over the big flag on the main screen. So that map generation can be reproduced even when no initial seeds were set in CivilizationIV.ini -- by reading the randomly generated seeds off the flag button and entering them in CivilizationIV.ini. This could become helpful for dealing with user feedback on the starting position iteration algorithm.

Changed the normalization code so that only map rand is used; in BtS, it's a mix of map rand and sync rand, and that gets in the way of reproducing regenerated maps.

And a small tweak to [advc.108]: Jungle on a resource in the ring of tiles surrounding a starting city radius is now cleared with a probability of 1/2 instead of 2/3.