Prefer starting sites with a resource just out of reach
Only when StartingPositionIteration (SPI) is enabled. Increase the found-city value a bit when there is a resource or other high-yield tile just outside the city radius. (This doesn't by any means guarantee that every player will find such a resource near their starting site.) In order to make the decision whether to move the starting settler more interesting. It's the first decision in the game and SPI had made settling in place optimal most of the time. [advc.027]