Handle assignment of random civs and leaders in the DLL
That is, keep track of external calls to CvRandom to determine which players receive a random civ or leader from the EXE. Display that information on the Settings tab, and use it to avoid giving away the – randomly chosen – civs and leaders of human rivals via the scoreboard in network games. [advc.190c]

(Some other, older additions to the Settings tab are now tagged with advc.190a, 190b instead of 004.)

Overwrite the random choices of the EXE so that civs with multiple leaders have a higher probability of being selected. (But not high enough to give every leader equal probability on the bottom line; Mansa Musa will still be chosen more often than Roosevelt.) [advc.191]