Refactor AI strength memory
[change id advc.158]

Moved into a separate class. Enabled for humans (and minor civs) because I think the only potential OOS issues with humans are in Debug-mode help text (fixed by this commit; change advc.001n) and in the cost function of the pathfinder (already addressed by karadoc). I'll still have to test it in multiplayer though. Strength memory disabled for Barbarians because I don't think Barbarians need to be able to bypass strongly defended tiles (and I've other plans for Barbarian strength memory).

Now uses a hash map, which was slightly faster than a vector in the test I've run.