Include groups in unit cycle when at least one unit ready
Alt. implementation of SHA
9806f41c71164910725e2cfa50db31650d00dd6d

I'm going to stick with the old implementation because it's more localized and allows certain missions (like route-to) to be excluded, but I still wanted to preserve this alternative approach.

I had hoped that I could let the game select the units that are able to move without breaking up the selection group stored internally. This works (and couldn't work with the old implementation, which needs to be able to split groups that aren't selected), but as soon as the player orders the selected units to skip their turn, they get stored as a separate group. I guess it should be possible to maintain or restore the old group in this special case, but, at least for the time being, this is getting too laborious for me.

Related CFC discussion:
https://forums.civfanatics.com/threads/advanced-civ.614217/page-46#post-16005262