Disable RMB back-cycling via worker cycling button
[change id advc.154]

When right-clicking the worker cycling button, the player would expect to go backwards in the worker cycle; however, such a worker cycling action isn't implemented. Best to disable the right-mouse interaction for the worker button (rather than let it go backwards in the regular unit cycle).

Another small amendment: Make sure not to display different unit icons on the worker and non-worker cycling buttons when they both refer to the same group.