Fix problems with improvement upgrade rate
Fixed a K-Mod bug that had caused the AI to value cottages /less/ highly when in Emancipation. (At least I'm pretty sure that this was wrong. Let's hope.) [advc.001]

Count the improvement upgrade rate at times-100 precision so that the Emancipation ability works as expected for values that aren't a multiple of 100 and for negative values (so long as they're not less than -100, i.e. still can't let Hamlets decay back to Cottages). [advc.912f]

To make this work, the number of turns that an improvement has been worked is now also counted at times-100 precision. This requires an int counter instead of a short int, but I've kept the size of CvPlot (kind of important for cache performance) the same as before by replacing another int with a char.