Rewrite output of combat modifiers in combat odds help
So that modifiers with an unexpected sign (e.g. -25% city attack) are handled correctly, i.e. shown with the correct sign and in the correct color. In AdvCiv, this had been a problem only for the Disorganized promotion (from K-Mod), but it's nice, for mod-mods, to have a flexible program logic. I've also eliminated redundancies, making the code easier to extend.