Fix GameFont issue (amendment)
Amends
e01fab22d13122f1f341efd79b32ff5dfab3a49e

It seems that the particular number of symbols and assignment to cells in that commit prevents the friendly-attitude smiley from being loaded. I hear that the GameFont file can be weird like that. I've now thrown out a duplicate of the cautious-attitude smiley -- which, I think, was intended as a generic attitude icon and, as far as I can tell, was unused. [advc.187]