Separate AgentIterator types for AI and non-AI civs
So that AI headers don't need to be included in non-AI code that uses the iterators. This hadn't been a problem so far because many of the non-AI implementation files need to include the AI headers in any case, - but this isn't and shouldn't be a given. Also, I had originally assumed that the agent instance obtained from an iterator would usually be assigned to some kLoopPlayer variable in the body of the loop, so that a distinction between AI and non-AI code would be made at that point (i.e. by giving kLoopPlayer type CvPlayerAI& only in AI code and CvPlayer& otherwise), but I've come to prefer using the iterator directly.