Earlier extra AI city defender on highest difficulty levels
Around the start of the Medieval era instead of Renaissance. Sort of in response to item 2 in this CFC post:
https://forums.civfanatics.com/threads/beating-advciv-diety-screenshots-discussion.668187/#post-16051984
I don't think this change goes far enough to help the AI in the situation described (Classical era?). Generally, considering that the AI unit production discounts are smaller in AdvCiv, I think the AI is better off not wasting early-game production on defenders for minor non-border cities, even if that means that human players can sometimes sneak up on those cities (which sounds like a fun stratagem too). [advc.107]