City defense bonuses only in territory of war allies
When a unit is attacked in a city or fort owned by a third party, the unit is a city defender only if the third party is also at war with the attacker, i.e. city defense and attack modifiers don't apply if the city owner is neutral (rationale: because the neutral city is not being attacked). [advc.183]

BtS had allowed city defense in neutral territory except for forts. In AdvCiv I had, by accident, already disabled some of the defensive modifiers in cities (regardless of the at-war status of the third party); so this AdvCiv bug should be fixed now.