paratrooper is now displayed available at flight in tech tree This is because flight is more intuitively linked to paratrooping.
Plus i mean the tech cost of Flight is almost double the cost of
Fascism so the Flight tech is even more preponderant this way.

Put it another way, it would be weird if after having researched
the Fascism tech, the player would then notice that paratroopers
are not available unlike what is displayed at first glance in
tech tree view. Same as the tweak(/fix) with the cavalry moved
to rifling in tech tree view now.

Functionally still the same unit.