add python and libboost gitignore so that i don't have to recommit this change every time i start a new PR (to hopefully better separate my "contributions" i mean anyways in several chunks).
For now i did not find where to store these outside of
AdvCiv folder, else they are not detected.
I assume there is a way to do it if they are not in
the gitignore of advciv, but for now works so is
convenient to just add these to the gitignore i
mean anyways.